pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether"
version = "0.1.0"
description = "Autonomous Carbon-Cost Arbitrage Engine for AI Compute"
authors = [{name = "AETHER Team"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    # Models
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    # Scheduling
    "apscheduler>=3.10.4",
    # Forecasting
    "prophet>=1.1.5",
    "scikit-learn>=1.4.2",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    # Optimization
    "scipy>=1.13.0",
    # Async file I/O
    "aiofiles>=23.2.1",
    # Cloud SDKs (all optional – mock fallback used when creds absent)
    "boto3>=1.34.0",
    "azure-mgmt-compute>=31.0.0",
    "azure-identity>=1.16.0",
    # Dashboard
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
    # Utilities
    "structlog>=24.1.0",
    "tenacity>=8.3.0",
    "python-dotenv>=1.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[project.scripts]
econode-api    = "econode.api.main:run"
econode-sched  = "econode.scheduler:run_standalone"

[tool.setuptools.packages.find]
where = ["."]
include = ["econode*", "dashboard*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA", "PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]
ignore = ["S101", "S104", "COM812", "ISC001"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
